updated architecture (closes #1)
