+++ documentationDEV.md
@@ -9,4 +9,4 @@ We will also split index.js into two different files:
 - api.js: it's the website's API
 - run.js: used to start the website
 
-Finally we will rename inMemoryWorkshop to db because it acts as a database.
+Finally we will rename db to db because it acts as a database.
