+++ src/front/templates/workshop.ejs
